requirements.txt
openai>=1.68.2
Pillow>=10.0.0
Kivy>=2.3.0
